update decks and user routes

+++ routes/api/decks.js
@@ -4,6 +4,7 @@ const mongoose = require('mongoose');
 const passport = require('passport');
 
 const Deck = require('../../models/Deck');
+const User = require('../../models/User');
 const validateDeckInput = require('../../validation/deck');
 
 // returns all public decks
@@ -14,17 +15,23 @@ router.get('/', (req, res) => {
     .catch(err => res.status(404).json({ nodecksfound: 'No decks found' }));
 });
 
-// returns specific deck
-router.get('/:id', (req, res) => {
-  Deck.findOne({ _id: req.params.id})
-    .then(deck => res.json(deck))
-    .catch(err =>
-      res.status(404).json({ nodeckfound: 'No deck found with that ID' })
-    );
+// returns specific deck. allows only owner or if desk if public
+router.get('/:id', passport.authenticate('jwt', { session: false }), async (req, res) => {
+  const deck = await Deck.findOne({ _id: req.params.id })
+  if (!deck) return res.status(404).json({ nodecksfound: 'No decks found with that ID' });
+  
+  const deckUser = await User.findOne({ _id: deck.user })
+
+  if (deckUser.id === req.user.id || deck.public) {
+    return res.json(deck);
+  } else {
+    return res.status(404).json({ nopermission: 'You do not have permission to view this deck' })
+  }
 });
 
 // returns all decks created by user
 router.get('/user/:user_id', (req, res) => {
+
   Deck.find({ user: req.params.user_id })
     .then(decks => res.json(decks))
     .catch(err =>
@@ -57,45 +64,51 @@ router.post('/',
 // edit deck owned by user
 router.patch('/:id',
   passport.authenticate('jwt', { session: false }),
-  (req, res) => {
+  async (req, res) => {
     const { errors, isValid } = validateDeckInput(req.body);
 
-    if (!isValid) {
-      return res.status(400).json(errors);
-    }
+    if (!isValid) return res.status(400).json(errors);
 
-    Deck.findOne({_id: req.params.id})
-      .then ( deck => {
-        const userDeck = deck.findOne({user: req.user.id});
-        console.log ('ud', userDeck);
-        console.log('id1', deck.findOne({ user: 1}));
+    const deck = await Deck.findOne({_id: req.params.id})
+
+    if (deck) {
+
+      const deckUser = await User.findOne({_id: deck.user})
+
+      if (deckUser.id === req.user.id) {
         deck.name = req.body.name;
         deck.category = req.body.category.split(',').map(cat => cat.trim());
         deck.public = req.body.public;
 
         deck.save().then(deck => res.json(deck))
-      })
-      .catch(err => res.status(404).json({ nodecksfound: 'No decks found from that user'}))
-
-    // console.log('req', req.user)
-    // console.log('user', req.user.id)
-    // console.log('deck', deck.user.id)
-    // console.log('deck', req.user.id !== deck.user.id)
-
-    // if (req.user.id !== deck.user) {
-    //   return res.status(400).json({invaliduser: 'This deck does not belong to you'});
-    // }
-   
-  
+      } else {
+        return res.status(404).json({ invaliduser: 'You do not own this deck' })
+      }
+    } else {
+      return res.status(404).json({ nodecksfound: 'No decks found with that id' });
+    }
   }
 );
 
 router.delete('/:id',
   passport.authenticate('jwt', { session: false }),
-  (req, res) => {
-    Deck.deleteOne({ _id: req.params.id })
-      .then(() => res.json({ deleted: "Deck was deleted"}))
-      .catch(err => res.status(404).json({ nodeckfound: 'No deck found with that ID' }))
+  async (req, res) => {
+    const deck = await Deck.findOne({ _id: req.params.id })
+
+    if (deck) {
+      const deckUser = await User.findOne({ _id: deck.user })
+
+      if (deckUser.id === req.user.id) {
+        Deck.deleteOne({ _id: req.params.id })
+        .then(() => res.json({ deleted: "Deck was deleted" }))
+        .catch(err => res.status(404).json({ nodeckfound: 'No deck found with that ID' }))
+      } else {
+        return res.status(404).json({ invaliduser: 'You do not own this deck' });
+      }
+      
+    } else {
+      return res.status(404).json({ nodeckfound: 'No deck found with that ID' })
+    }
   }
 );
